std: Clarify invalidation in HashMap documentation
This commit makes an effort to reduce mistakes made in writing code
utilising Zig's hashmap.

Previously, the fact that modifying the hash map has the potential to
invalidate pointers to entries in the hash map was not documented, and
may not have been immediately obvious to a developer. This has the
potential to lead to difficult-to-reproduce segfaults, or illegal
behaviour.

Additionally, functions which can modify the hash map were not
documented as having the potential to do so. This could lead to
ambiguity, and thus mistakes, in understanding which functions will
invalidate pointers and iterators, and which will not.

TODO: Other hash maps